Give advice to change to runtimeOnly where appropriate
If an advice was computed to change a declaration from
'implementation'/'api' to 'runtimeOnly', this is indeed correct
information. The dependency is *not* required at compile time.
But it *might* be required at runtime.

It is not possible for this plugin to make a clear statement about
if it is actually needed at runtime. When getting such an advice, the
user might also remove the dependency completely. The 'reason' feature
will give more insight on why the dependency *might* be needed at
runtime. The following cases are covered:

- Contains Service Loader declarations
- Contains Security Provider declarations
- Contains Content Provider or Service declarations (Android)
- Contains Android linters (Lint-Registry entry)