Put 'compileOnlyApi' and 'providedCompile' into the COMPILE_ONLY bucket
From the perspective of this plugin right now, dependencies defined as
'compileOnlyApi' or 'providedCompile' should be treated the same as
'compileOnly'. Which is, there are no advices produced for them.

In the future, the plugin might make suggestion whether a 'compileOnly'
dependency is "api" or not, but right now such a feature is not
implemented.

The 'providedCompile' configuration is added by the 'war' Gradle core
plugin if applied. This configuration exists for years already. If
it would be introduced today, it would probably be called
'providedApi'. It has the same behaviour as the 'api' configuration,
but with the difference that it is not packaged into the 'war' file
that makes up the final application. Because the Jar behind the expected
dependency is expected to be provided by the runtime environment.
Currently, there is no 'providedImplementation'. So an analysis whether
a 'providedCompile' dependency has "api" or "implementation"
characteristics won't be useful because there is not distinction
right now.