Support Components with multiple Jar files
In Gradle, one component - i.e. one logical component referred to
by GA coordinates or Gradle project path - can have multiple Jars.
So far, the plugin would silently ignore all Jars except for the first
it finds. With this fix, all Jars are considered part of the component
during the analysis.
(This used to work in older versions; tested with 0.71.0)