Support Components with multiple Jar files (#688)
* Clarify comments about wrong advice when Capabilities are involved

* Add test case for a Component with multiple Jars

* Add another test case with a simpler project setup

* Adjust test for behavior that is now 'more correct'

* Remove duplicated repository declaration

* Support Components with multiple Jar files

In Gradle, one component - i.e. one logical component referred to
by GA coordinates or Gradle project path - can have multiple Jars.
So far, the plugin would silently ignore all Jars except for the first
it finds. With this fix, all Jars are considered part of the component
during the analysis.
(This used to work in older versions; tested with 0.71.0)