Fixes #763
If an annotation references multiple types they are not correctly exposed as part of a module's public ABI. This is because we filter for an annotation values to only include a single `Type`. However annotations may have a value of type `ArrayList<*>` which can contain references to other classes.

This patch updates the `List<AnnotationNode>?.annotationTypes()` method to correctly handle the above scenario and adds some unit tests for it.