Duplicated declaration does not lead to wrong analysis result
If there are two declarations to the same component with different
scopes, e.g.

runtimeOnly("org.apache.commons:commons-collections4:4.4")
implementation("org.apache.commons:commons-collections4:4.4")

the analysis discarded one of them by using 'find' instead of
'filter' in StandardTransform.kt. This can lead to a wrong
recommendation. In the example provided in the test, the plugin
recommended to remove the 'implementation' dependency.

A related feature could be that the plugin also detects the duplication,
and recommends to remove the unneeded declaration - in this case
runtimeOnly("..."). Doing this, would be an additional analysis imo,
as the plugin would need to check how the different configurations
relate. And maybe this is also not wanted in all cases. The setup
is not "wrong" as the classpaths don't change by removing the
duplicates. It may be out of scope of this plugin altogether.

This fix is only about not reporting something wrong.