The requested version is irrelevant for included builds
When the version is in the coordinates, it is used as part of the
identity (e.g. in the dependency json file name). Then different
versions look like different components in the analysis. But Gradle
resolves ALL requested versions to the included build. I.e., if
there is an included build matching the GA coordinates, the
requested version is ignored completely.