Analyze dependencies of all source sets in standard JVM projects
This adds analyzing dependencies FROM other source sets (which are not
'main' or 'test') in standard Gradle 'java-library' and 'application'
projects.

For example:
- 'testFixturesImplementation(...)'
- 'integrationTestImplementation(...)'

This is solved as follows:

- All source sets registered in the 'SourceSetContainer' are marked
  as supported for 'java-library' and 'application' projects.
- For each source set, we check if it can have an ABI by checking if
  the 'api...' configuration was created. This happens, if you define
  a Feature Variant on top of a source set.
- There is a new SourceSetKind.CUSTOM_JVM that is used for all
  source sets which are not 'main' or 'test'. In this case, the internal
  'base variant' object is different (different name) depending on the
  source set's name.

At first, I considered having only one SourceSetKind for all JVM
source sets (including 'main' and 'test'). But there is still the
special behavior that main dependencies are automatically seen by test.

I think it is debatable, if this is really the right behavior.
I.e., if I use something directly in my tests, why not also have a
'testImplementation' dependency in addition to 'implementation'.
But we probably don't want to change the current behavior. It could
be reconsidered in the future though and then I think all source sets
in JVM projects could be treated equally (have the same SourceSetKind).

For custom source sets, I think the 'stricter' behavior is right,
as they usually can't automatically see dependencies of other source
sets.