Add test to showcase included build annotation processor issue
We are seeing issues with using annotation processor results from
included builds when retrieving the information from the cache.
I mirrored our setup in this test, by introducing two new projects:
1. A new included build called `processor-build` that defines an
annotation processor and uses the Google AutoService to define it
2. A new submodule in this project, which uses the processor to
process one of its fields

The problem is the following (although I couldn't fully define that
in the test): when we run `:projectHealth` on this, it stores the information
in our build cache. When we then attempt to retrieve it in a subsequent
run, the `:projectHealth` task fails with the following error:

```
> Transitively used dependencies that should be declared directly as indicated:
    annotationProcessor project('my.custom.processor:sub-processor1')
```

It seems like the information retrieved from the cache is incomplete,
where it somehow deduces that the annotation processor is unused.
Again, if we run it without the cache, it is all fine. Only when
the `:user-of-processor:findDeclaredProcsMain` is `FROM-CACHE`,
it fails.

This only happens when using included builds. If the subprojects are
part of the same build, we don't observe any caching problems.

Next to that, we also noticed that the `usages-annotation-processors.json`
includes fully qualified project references. However, the cache key
for this task does not take into account the fully qualified names
of dependencies. In the sense that: when we run the task in 1 build
(where the `rootProject.group` is set to `Group1`) and we run it again
in another build where the group is set to `Group2`, the task would also
fail, as it doesn't match the group.

We fixed this issue by aligning the group names, but I think that's
a problem that can be fixed in the plugin as well. To do so, we can
mark the `rootProject.group` as `inputs.property()` to these tasks.
(Note that the `rootProject.group` is of the dependency, not the project
that defines the plugin task, which makes it slightly more complicated).