Consistently use IncludedBuildCoordinates to refer to other projects
There is a bug that leads to wrong/inconsistent advices if included
builds are involved. The assumption at the moment is that if a
'project path' matches the 'identity path', the dependency is not in
an included build.

This is in:
`gradleStrings.ResolvedArtifactResult.toCoordinates()`

While this works in the case where we do not use included builds (much),
it is generally wrong. There can also be references to other projects
inside other included builds. And, more problematic, an included
build can refer back to the "root" build. The plugin then assumes
that a a project in the root build could be addressed via
'project(":path")' notation from that other build, which is not true.

This is an attempt to use IncludedBuildCoordinates consistently
throughout the analysis. Then we also do not need to handle both
in the middle of the advice calculation. Then only in the end when
an advice is given, we decide which representation of the coordinates
is shown to the user. For 'add' advices, we need to track the build ID
(build name), to decide if a project is in the same build and can
hence be addressed by project coordinates.