fix: handle disjoint classpaths.
If the main source set has a used dependency, and a custom source set has an unused transitive dependency,
and these are at different versions, StandardTransform was advising the removal of the used dependency.
Now we check that the usage is not undeclared before adding to the set of remove-advice.