fix: ugly workaround for issue 1116.
Basic hypothesis is that tasks.whenTaskAdded eagerly realizes all tasks, which forces DAGP tasks to be
configured too soon. Some of them have dependencies on AGP tasks that are registered later in the build
lifecycle. This workaround uses afterEvaluate as a bandaid. Preferable would be using some Android-
specific version of SourceSetContainer to get lazy access to compiled class files, instead of relying
on tasks directly.