feat: detect duplicate classfiles on compile and runtime classpaths, and warn.
At this point, not sure what to do but warn. Providing more detailed advice would
require much more information about the _structure_ of the class files, which is
a problem we've avoided till now. We'd basically need an AST and an understanding
of how the classes are used, vs simply seeing the classes referenced in bytecode.