fix: use projectPath, not identityPath, for included build coordinates.
This has been a longstanding issue and will probably have multiple knock-on effects.
The immediate reason I saw this was that SuperGraph was trying to find nodes built with one version of a graph using nodes from
another version of a graph, and failing. This change should sync the two, which ought to also mean fewer outputs in the
dependencies directory (no duplication for included build dependencies). I think lots of other code was working around this problem
too without realizing it.